# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: vrp-0 API
  version: 1.0.4-alpha-SNAPSHOT
...
paths:
...
  /solver_job:
    get:
      tags:
      - VRP求解引擎
      summary: 获取当前求解任务
      parameters:
      - name: remove_virtual
        in: query
        description: 是否清除虚拟工程师指派
        schema:
          default: false
          type: boolean
      responses:
        "200":
          description: 当前求解任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolverJob'
        "404":
          description: 当前没有求解任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
    post:
      tags:
      - VRP求解引擎
      summary: 基于当前场景提交求解
      parameters:
      - name: build_transit_matrix
        in: query
        description: 记录到任务并用于列表筛选的矩阵构建选项；当前求解阶段不会重新构建输入矩阵
        schema:
          default: false
          type: boolean
      - name: callback
        in: query
        description: 求解完成回调URL
        schema:
          default: ""
          type: string
      - name: draw_route
        in: query
        description: 是否生成完整路线规划
        schema:
          type: boolean
      - name: matrix_mode
        in: query
        description: 记录到任务的矩阵方式元数据：ROUTING、MANHATTAN；当前求解要求输入矩阵已预先构建，AMAP 仅兼容旧调用
        schema:
          allOf:
          - $ref: '#/components/schemas/MatrixMode'
          - default: ROUTING
      - name: solve_time
        in: query
        description: 期望求解时间，ISO 8601 duration format
        schema:
          default: PT30S
          type: string
      responses:
        "200":
          description: 已提交的当前求解任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolverJob'
        "404":
          description: 当前没有场景
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
        "409":
          description: 已有运行中的求解任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
    delete:
      tags:
      - VRP求解引擎
      summary: 删除当前求解任务
      responses:
        "200":
          description: 请求成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Msg'
        "404":
          description: 当前没有求解任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
  /solver_job/apply:
    post:
      tags:
      - VRP求解引擎
      summary: 将当前求解结果应用到当前场景
      responses:
        "200":
          description: 应用后的当前场景
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
        "404":
          description: 当前没有场景或求解任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
  /solver_job/list:
    get:
      tags:
      - VRP求解引擎
      summary: 按条件获取求解任务历史列表
      description: 未传筛选条件时返回全部任务，并按创建时间倒序排列。创建时间范围使用 ISO 本地日期时间，首尾均包含。
      parameters:
      - name: build_transit_matrix
        in: query
        description: 按任务记录中的矩阵构建选项筛选
        schema:
          type: boolean
      - name: create_time_from
        in: query
        description: 创建时间起点（ISO 本地日期时间，包含）
        schema:
          pattern: yyyy-MM-dd'T'HH:mm:ss
          type: string
      - name: create_time_to
        in: query
        description: 创建时间终点（ISO 本地日期时间，包含）
        schema:
          pattern: yyyy-MM-dd'T'HH:mm:ss
          type: string
      - name: draw_route
        in: query
        description: 是否生成完整路线规划
        schema:
          type: boolean
      - name: matrix_mode
        in: query
        description: 按任务记录中的矩阵方式筛选：ROUTING、MANHATTAN；AMAP 兼容旧记录
        schema:
          $ref: '#/components/schemas/MatrixMode'
      - name: status
        in: query
        description: 任务状态
        schema:
          $ref: '#/components/schemas/Status1'
      responses:
        "400":
          description: 筛选参数非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
        "200":
          description: 求解任务摘要列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SolverJobSummary'
  /solver_job/terminate:
    post:
      tags:
      - VRP求解引擎
      summary: 终止当前求解任务
      responses:
        "200":
          description: 当前求解任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolverJob'
        "404":
          description: 当前没有求解任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInfo'
  /solver_job/{id}:
    get:
      tags:
      - VRP求解引擎
      summary: 按任务ID获取求解任务
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: remove_virtual
        in: query
        description: 是否清除虚拟工程师指派
        schema:
          default: false
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolverJob'
    delete:
      tags:
      - VRP求解引擎
      summary: 删除指定求解任务
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Msg'
  /solver_job/{id}/apply:
    post:
      tags:
      - VRP求解引擎
      summary: 将指定求解结果应用到当前场景
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scenario'
components:
  schemas:
...
    CostParameter:
      type: object
      properties:
        start_price:
          format: double
          description: 起步价
          type: number
        node_fee:
          format: double
          description: 节点费
          type: number
        max_overload_ratio:
          format: double
          description: 最大超载比例
          type: number
        overload_fee:
          format: double
          description: 超载费
          type: number
        cross_region_threshold:
          format: double
          description: 跨区阈值 (米)
          type: number
        cross_region_fee:
          format: double
          description: 跨区费
          type: number
        guaranteed_income:
          format: double
          description: 保底费，单位元/车次
          type: number
        time_restricted_traffic_charge:
          format: double
          description: 限行时段内行驶的额外费用
          type: number
        elec_price:
          format: double
          description: 电费
          type: number
        gas_92_price:
          format: double
          description: 92油费
          type: number
    Depo:
      required:
      - id
      type: object
      properties:
        loc:
          description: 位置。可传 plan.pois 中 POI 的 ID 字符串，或直接传 POI 对象。
          example: B0G2X7N5D2
          oneOf:
          - type: string
          - $ref: '#/components/schemas/POI'
        id:
          description: 仓库id
          type: string
        name:
          description: 仓库名称
          type: string
    Duration:
      format: duration
      type: string
      example: P1D
    ErrorInfo:
      description: 统一 REST 错误响应。网页客户端使用 error_code 和 error_params 本地化展示，message 仅保留给兼容客户端和诊断。
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        error_code:
          description: 稳定、面向机器的错误码
          type: string
        error_params:
          description: 错误码插值所需的非敏感参数
          type: object
          additionalProperties:
            type: string
        message:
          description: 向后兼容的服务端诊断消息；网页不得直接展示
          type: string
...
    ExtendedStackTraceElement:
      type: object
      properties:
        extraClassInfo:
          $ref: '#/components/schemas/ExtendedClassInfo'
        stackTraceElement:
          type: object
        className:
          type: string
        exact:
          type: boolean
        fileName:
          type: string
        lineNumber:
          format: int32
          type: integer
        location:
          type: string
        methodName:
          type: string
        version:
          type: string
        nativeMethod:
          type: boolean
...
    LocalDate:
      format: date
      type: string
      example: 2022-03-10
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
...
    MapProvider:
      description: 地图服务提供商
      enum:
      - AMAP
      - HERE
      type: string
    MatrixMode:
      enum:
      - ROUTING
      - MANHATTAN
      - AMAP
      type: string
...
    Metrics:
      type: object
      properties:
        distance_total:
          format: int64
          type: integer
        duration_total:
          format: int64
          type: integer
        ton_total:
          format: double
          type: number
        cost_total:
          format: double
          type: number
        cost_per_ton_per_km:
          format: double
          type: number
        score:
          type: string
          example: 0hard/-5medium/-79909soft
    Msg:
      type: object
      properties:
        code:
          format: int32
          type: integer
        msg:
          type: string
    POI:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        location:
          description: 经纬度
          type: string
        loc:
          $ref: '#/components/schemas/LOC'
        address:
          description: 详细地址
          type: string
        pcode:
          description: 所属省份编码
          type: string
        type:
          description: 所属类型
          type: string
        photos:
          description: ' poi 图片相关信息'
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        gridcode:
          description: ' poi 的地理格 id'
          type: string
        typecode:
          description: 分类编码
          type: string
        citycode:
          description: 所属城市编码
          type: string
        adname:
          description: 所属区域名称
          type: string
        entr_location:
          description: 入口经纬度坐标
          type: string
        entr_loc:
          $ref: '#/components/schemas/LOC'
        adcode:
          description: 所属区域编码
          type: string
        pname:
          description: 所属省份名称
          type: string
        cityname:
          description: 所属城市名称
          type: string
...
    RecordType:
      enum:
      - BEST_SOLUTION
      - SEARCH_SAMPLE
      - FINAL_SOLUTION
      type: string
...
    RoutePlan:
      required:
      - depos
      - agents
      - tickets
      type: object
      properties:
        skus:
          type: array
          items:
            $ref: '#/components/schemas/SKU'
        pois:
          description: 地址信息
          type: array
          items:
            $ref: '#/components/schemas/POI'
        depos:
          description: 仓库地址
          type: array
          items:
            $ref: '#/components/schemas/Depo'
        agents:
          description: 工程师/车辆列表
          type: array
          items:
            $ref: '#/components/schemas/AgentEachDay'
        tickets:
          description: 工单列表
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        matrix:
          description: 迁移矩阵
          type: object
          allOf:
          - $ref: '#/components/schemas/TransitMatrix'
        constraint_configuration:
          description: 约束配置项
          type: object
          allOf:
          - $ref: '#/components/schemas/RoutePlanConstraintConfiguration'
        cost_parameter:
          description: 成本参数
          type: object
          allOf:
          - $ref: '#/components/schemas/CostParameter'
        score:
          description: 规划评分
          type: string
          example: 0hard/-5medium/-79909soft
        score_explanation:
          description: 求解解释，内部字段，不需要传参
          type: string
        metrics:
          type: object
          allOf:
          - $ref: '#/components/schemas/Metrics'
          readOnly: true
    RoutePlanConstraintConfiguration:
      type: object
      properties:
        name:
          type: string
        车辆装载容量限制:
          description: agent装载不能超过其设计上限，相关字段：agent.weight、agent.vol，默认惩罚：1 Hard
          type: string
        最大接单:
          description: agent最大接单量不应大于设置值，相关字段：agent.max_ticket_num，默认惩罚：1 Hard
          type: string
        标签匹配:
          description: 所指派agent的标签包含ticket标签，相关字段：agent.skills、ticket.skillsRequired，默认惩罚：1
            Hard
          type: string
        需求技能等级匹配:
          description: agent等级与ticket等级相匹配，相关字段：agent.qualificationLevels、ticket.qualificationLevelsRequired，默认惩罚：100
            medium
          type: string
        工单时序因果依赖:
          description: 下一级ticket的完成时间必须晚于上级ticket，相关字段：ticket.dep_tickets、ticket.ref_tickets，默认惩罚：0
            Hard（未启用）
          type: string
        工单工程师依赖:
          description: 下一级ticket与上级ticket必须分配给同一agent，相关字段：ticket.dep_tickets、ticket.ref_tickets、ticket.agent_id，默认惩罚：0
            Hard（未启用）
          type: string
        工单截止时间:
          description: ticket完成不能晚于最后截止时间，相关字段：ticket.max_end_time，默认惩罚：0 Hard（未启用）
          type: string
        时间窗约束:
          description: ticket在客户预期时间上门，相关字段：ticket.min_start_time、ticket.max_end_time，默认惩罚：100
            Soft（每分钟）
          type: string
        当日指派:
          description: ticket必须在当日指派工程师，相关字段：ticket.create_time，默认惩罚：1 Hard
          type: string
        关联工单相同指派:
          description: ticket与其关联ticket尽量派发给相同工程师，相关字段：ticket.relation_tickets，默认惩罚：50
            Medium
          type: string
        最小行驶时间:
          description: 所有agent的总在途时间尽可能小，默认惩罚：1 Soft（每分钟）
          type: string
        最化行驶距离:
          description: 所有agent的总在途里程尽可能小，默认惩罚：0 Soft（每米）（未启用）
          type: string
        最小固定成本:
          description: 总固定成本最小，默认惩罚：20 Soft（每元）
          type: string
        工单工程师同网点:
          description: ticket只能指派给同网点的agent，相关字段：agent.depo_id、ticket.depo_id，默认惩罚：1
            Hard
          type: string
        工单负载均衡:
          description: 分配给每个agent的ticket应差异不大，相关字段：agent.tickets，默认惩罚：1 Medium
          type: string
        负载比例均衡:
          description: 每个agent的有效载荷使用比例应该差异不大，默认惩罚：0 Medium（未启用）
          type: string
        工作时长负载均衡:
          description: 每个agent的预期工作时长应该差异不大，默认惩罚：0 Medium（未启用）
          type: string
        最小工单变更:
          description: 尽量不要改派已经分配的工单，默认惩罚：1000 Soft
          type: string
        虚拟工程师:
          description: 超约束优化时使用，需要权衡业务场景和客户需求设计惩罚值。如果违背时间窗约束带来的惩罚大于虚拟工程师的惩罚，求解时倾向于虚拟工程师指派，也就是说该工单不进行配送，反之亦然。默认惩罚：1000
            Medium
          type: string
...
    SKU:
      required:
      - id
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        weight:
          format: float
          description: 重量
          type: number
        vol:
          format: float
          description: 体积
          type: number
    Scenario:
      required:
      - name
      - planning_date
      - start_time
      - end_time
      - plan
      type: object
      properties:
        id:
          type: string
          allOf:
          - $ref: '#/components/schemas/UUID'
          writeOnly: true
        name:
          description: 场景名称
          type: string
        desc:
          description: 场景描述
          type: string
        map_provider:
          description: 场景坐标及地图能力使用的图商；缺省时由当前 MAP_PROVIDER 补齐
          enum:
          - AMAP
          - HERE
          type: string
          allOf:
          - $ref: '#/components/schemas/MapProvider'
        planning_date:
          description: 当前规划日期
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDate'
        start_time:
          description: 规划开始日期
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
        end_time:
          description: 规划结束日期
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
        plan:
          description: 规划方案
          type: object
          allOf:
          - $ref: '#/components/schemas/RoutePlan'
        create_time:
          description: 场景创建时间
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
        update_time:
          description: 场景更新时间
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
        poi_build:
          type: boolean
          readOnly: true
        matrix_build:
          type: boolean
          readOnly: true
    SolutionMetrics:
      description: 求解过程的真实 Best、搜索快照或最终解记录
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        metrics:
          $ref: '#/components/schemas/Metrics'
        create_time:
          $ref: '#/components/schemas/LocalDateTime'
        recorded_at_millis:
          format: int64
          description: 记录产生时的 Unix 毫秒时间戳
          type: integer
        elapsed_millis:
          format: int64
          description: 从本次实际搜索开始到记录时刻的毫秒数
          type: integer
        search_progress:
          description: 仅 record_type=SEARCH_SAMPLE 时存在的搜索快照
          type: object
          allOf:
          - $ref: '#/components/schemas/SolverSearchProgress'
        record_type:
          description: BEST_SOLUTION 仅在得分严格改善时产生；SEARCH_SAMPLE 为至多每秒一次的搜索快照；FINAL_SOLUTION
            使用搜索结束时刻
          type: string
          allOf:
          - $ref: '#/components/schemas/RecordType'
    SolverJob:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        scenario_name:
          description: 场景名称
          type: string
        name:
          description: 路径规划方案名称
          type: string
        start_date_time:
          description: 规划开始时间
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
        end_date_time:
          description: 规划结束时间
          type: string
          allOf:
          - $ref: '#/components/schemas/LocalDateTime'
        plan:
          description: 规划方案
          type: object
          allOf:
          - $ref: '#/components/schemas/RoutePlan'
        solve_time:
          description: 求解时长
          type: string
          allOf:
          - $ref: '#/components/schemas/Duration'
        matrix_mode:
          description: 任务提交时记录的矩阵方式元数据；当前求解阶段不会据此重建输入矩阵
          type: string
        build_transit_matrix:
          description: 任务提交时记录的矩阵构建选项元数据；当前求解阶段不会据此重建输入矩阵
          type: boolean
        draw_route:
          description: 是否生成完整路线规划
          type: boolean
        status:
          description: 求解状态
          type: string
          allOf:
          - $ref: '#/components/schemas/Status1'
        score_explanation:
          description: 求解解释
          type: string
        exception:
          description: 求解时发生的异常
          type: object
          allOf:
          - $ref: '#/components/schemas/ThrowableProxy'
        solution_metrics_list:
          type: array
          items:
            $ref: '#/components/schemas/SolutionMetrics'
        create_time:
          $ref: '#/components/schemas/LocalDateTime'
        update_time:
          $ref: '#/components/schemas/LocalDateTime'
        score:
          description: "求解得分，其中分为三个部分[hard/medium/soft]，如果hard为负，则认为没有可行解"
          type: string
          readOnly: true
          example: 0hard/-5medium/-79909soft
        cc_name:
          type: string
          readOnly: true
        constraint_configuration:
          type: object
          allOf:
          - $ref: '#/components/schemas/RoutePlanConstraintConfiguration'
          readOnly: true
        cost_parameter:
          type: object
          allOf:
          - $ref: '#/components/schemas/CostParameter'
          readOnly: true
        metrics:
          $ref: '#/components/schemas/Metrics'
    SolverJobSummary:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        scenario_name:
          type: string
        name:
          type: string
        status:
          $ref: '#/components/schemas/Status1'
        solve_time:
          $ref: '#/components/schemas/Duration'
        matrix_mode:
          type: string
        build_transit_matrix:
          type: boolean
        draw_route:
          type: boolean
        score:
          description: 当前最佳或最终求解得分，分为 hard/medium/soft 三部分
          type: string
          example: 0hard/-5medium/-79909soft
        create_time:
          $ref: '#/components/schemas/LocalDateTime'
        update_time:
          $ref: '#/components/schemas/LocalDateTime'
    SolverSearchProgress:
      description: 求解过程搜索快照
      type: object
      properties:
        current_score:
          description: 最近完成搜索步骤的得分
          type: string
          example: 0hard/-58medium/-4612soft
        best_score:
          description: 截至当前的历史最优得分
          type: string
          example: 0hard/-54medium/-4341soft
...
    Status1:
      enum:
      - SOLVING_SCHEDULED
      - SOLVING_ACTIVE
      - NOT_SOLVING
      - SOLVING_FINISHED
      - ERROR
      type: string
    ThrowableProxy:
      type: object
      properties:
        causeProxy:
          $ref: '#/components/schemas/ThrowableProxy'
        commonElementCount:
          format: int32
          type: integer
        extendedStackTrace:
          type: array
          items:
            $ref: '#/components/schemas/ExtendedStackTraceElement'
        localizedMessage:
          type: string
        message:
          type: string
        name:
          type: string
        suppressedProxies:
          type: array
          items:
            $ref: '#/components/schemas/ThrowableProxy'
        extendedStackTraceAsString:
          type: string
        stackTrace:
          type: array
          items:
            type: object
...
    TransitMatrix:
      type: object
      properties:
        data:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              $ref: '#/components/schemas/Transit'
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...